requirements.txt
qq-botpy
requests
pyyaml
schedule